app/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
werkzeug==3.0.0
flask-jwt-extended==4.6.0